+++ src/lib.rs
@@ -2,6 +2,7 @@ pub mod atom;
 pub mod backup;
 pub mod db;
 pub mod env;
+pub mod export;
 pub mod http;
 pub mod ics;
 pub mod json;